+++ pkg/common/src/util/network_utils.h
@@ -4,5 +4,5 @@
 #include <string>
 
 std::string get_local_ip();
-int get_available_port(uint32_t start_port = 10001, uint32_t max_port = 65535);
+uint16_t get_available_port(uint16_t start_port = 10001, uint16_t max_port = 65535);
